Add angular and small test app loaded via system.js
also add some basic gulp tasks and an eslint config file.

After a couple of hours of trying to get Angular 2 to work i gave up when i realised not even the the bulls scripts in the angular/quickstart repo on github works. I'll start the app with angular 1 and upgrade later once angular 2 is out of alpha.

Using system.js and es6-module-loader together with angular is working at last, but it doesn't quite feel like they fit together. And while the page loads dynamically, it loads really slow.