Add unittesting with karma.
This was hard to figure out. The gulp-jspm seems to really fit with my
use-case, since it doesn't follow the same paths as karma. In the end i
got it to work byt putting my angular files in a subfolder and proxying
the url in karma.

This did solve some trouble i had with file path
globbing which is positive.