Added angular-material
Installing angular-material with jspm did add a lot of dependencies but
using it in the app was super simple. It did however break the import of
the unit test but that was easily fixed once the problem was found.

Using a template language like jade is also very nice, not having to
bother with closing tags. It feels quite like python and I like it.